+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Thanks to @nate-sys and @jubalh for contributing to this release.
 - Default config updated with better comments, color config for join/part
   messages. (#412)
 - `/join` command errors now print usage help once instead of twice.
+- Fix showing timestamp of the next message or activity after a `/clear`.
+  (#417)
 
 # 2023/07/16: 0.11.0